Cleaned up tests folder, introduced integration/unit seperation
