+++ GraphGenerator/Tests/CMakeLists.txt
@@ -1,0 +1,17 @@
+set(FILES )
+set(FILES )
+
+set(SUBDIRECTORIES 
+	Integration
+	Units
+	Utility
+)
+
+foreach(VAR ${SUBDIRECTORIES})
+	target_include_directories(${PROJECT_NAME} PUBLIC ${VAR})
+	add_subdirectory(${VAR})
+endforeach()
+
+if(FILES)
+	target_sources(${PROJECT_NAME} PUBLIC ${FILES})
+endif()
