+++ GraphGenerator/Tests/Units/sample_empty.json
@@ -1,0 +1,8 @@
+{
+	"dependencies" : null,
+	"sources" : 
+	[
+		"free_functions.cpp"
+	],
+	"structures" : null
+}
