+++ GraphGenerator/Tests/Utility/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(FILES 
+	Utility.cpp 
+	Utility.h
+)
+set(SUBDIRECTORIES )
+
+foreach(VAR ${SUBDIRECTORIES})
+	target_include_directories(${PROJECT_NAME} PUBLIC ${VAR})
+	add_subdirectory(${VAR})
+endforeach()
+
+if(FILES)
+	target_sources(${PROJECT_NAME} PUBLIC ${FILES})
+endif()
